renames views folder to pages
