+++ src/pages/examples/Register.js
